:wastebasket: Cleaning up the codebase
