+++ src/functions/external/checkExternal.ts
@@ -1,0 +1,63 @@
+import { GoogleSafeBrowsing } from "./apis/GoogleSafeBrowsing";
+import { IpQualityScore } from "./apis/IpQualityScore";
+import { Phisherman } from "./apis/Phisherman";
+import { SinkingYahts } from "./apis/SinkingYahts";
+import { Walshy } from "./apis/Walshy";
+
+/**
+ *
+ * @param link
+ */
+export const checkExternal = async (link: string) => {
+  try {
+    // try different APIs
+    const PhishermanResponse = await Phisherman(`${link}`, true, false);
+    if (PhishermanResponse) {
+      return {
+        scamDetected: true,
+        source: "Phisherman",
+      };
+    }
+
+    const IpQualityScoreResponse = await IpQualityScore(`${link}`, true, false);
+    if (IpQualityScoreResponse) {
+      return {
+        scamDetected: true,
+        source: "IpQualityScore",
+      };
+    }
+    const WalshyResponse = await Walshy(`${link}`, true, false);
+    if (WalshyResponse) {
+      return {
+        scamDetected: true,
+        source: "Walshy",
+      };
+    }
+    const SinkingYahtsResponse = await SinkingYahts(`${link}`, true, false);
+    if (SinkingYahtsResponse) {
+      return {
+        scamDetected: true,
+        source: "SinkingYahts",
+      };
+    }
+    const GoogleSafeBrowsingResponse = await GoogleSafeBrowsing(
+      `${link}`,
+      true,
+      false
+    );
+    if (GoogleSafeBrowsingResponse) {
+      return {
+        scamDetected: true,
+        source: "GoogleSafeBrowsing",
+      };
+    }
+
+    // if no API detects a scam, return false
+    return {
+      scamDetected: false,
+      source: "",
+    };
+  } catch (err) {
+    console.error(err);
+  }
+};
