+++ src/functions/flattenLink.ts
