+++ src/functions/validateEnv.ts
