+++ src/index.ts
@@ -4,7 +4,7 @@ import mongoose from "mongoose";
 import "dotenv/config";
 
 import app from "./app";
-import { validateEnv } from "./utils/validateEnv";
+import { validateEnv } from "./functions/validateEnv";
 
 const PORT = process.env.PORT;
