--- src/utils/getLatitudeAndLongitude.ts
@@ -1,7 +1,0 @@
-export function getLatitudeAndLongidute(location: string) {
-  const [latitude, longitude] = location.split(",");
-  return {
-    latitude: latitude,
-    longitude: longitude,
-  };
-}
