--- src/utils/getLocationDetails.ts
@@ -1,10 +1,0 @@
-import { IMetrics } from "../models/Metrics";
-
-export function getLocationDetails(metrics: any): Partial<IMetrics> {
-  const locationInfo = {
-    city: metrics.city,
-    region: metrics.region,
-    country: metrics.country,
-  };
-  return locationInfo;
-}
